restructure the project and add babel so we can use es6
