+++ tests/index.spec.js
@@ -1,4 +1,4 @@
-const app = require('../app')
+const app = require('../src/app')
 const request = require('supertest')
 
 describe('GET /', () => {
